Merge protected area and carbon updates into master

+++ src/get_cell_area.c
@@ -124,7 +124,7 @@ int get_cell_area(args_struct in_args, rinfo_struct *raster_info) {
     }
     
     // read the data
-    num_read = fread(cell_area_hyde, insize, ncells, fpin);
+    num_read = (int) fread(cell_area_hyde, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != ncells)
     {
@@ -167,5 +167,4 @@ int get_cell_area(args_struct in_args, rinfo_struct *raster_info) {
 		}
 	}
 	
-	return OK;
-}
+	return OK;}
