+++ src/read_aez_new.c
@@ -89,7 +89,7 @@ int read_aez_new(args_struct in_args, rinfo_struct *raster_info) {
 	}
 	
 	// read the data
-	num_read = fread(aez_bounds_new, insize, ncells, fpin);
+	num_read = (int) fread(aez_bounds_new, insize, ncells, fpin);
 	fclose(fpin);
 	if(num_read != ncells)
 	{
@@ -105,5 +105,4 @@ int read_aez_new(args_struct in_args, rinfo_struct *raster_info) {
 		}
 	}
 	
-	return OK;
-}
+	return OK;}
