+++ src/read_aez_orig.c
@@ -90,7 +90,7 @@ int read_aez_orig(args_struct in_args, rinfo_struct *raster_info) {
 	}
 	
 	// read the data
-	num_read = fread(aez_bounds_orig, insize, ncells, fpin);
+	num_read = (int) fread(aez_bounds_orig, insize, ncells, fpin);
 	fclose(fpin);
 	if(num_read != ncells)
 	{
@@ -106,5 +106,4 @@ int read_aez_orig(args_struct in_args, rinfo_struct *raster_info) {
 		}
 	}
 	
-	return OK;
-}
+	return OK;}
