+++ src/read_potveg.c
@@ -88,7 +88,7 @@ int read_potveg(args_struct in_args, rinfo_struct *raster_info) {
     }
     
     // read the data
-    num_read = fread(potveg_thematic, insize, ncells, fpin);
+    num_read = (int) fread(potveg_thematic, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != ncells)
     {
@@ -104,5 +104,4 @@ int read_potveg(args_struct in_args, rinfo_struct *raster_info) {
 		}
 	}
 	
-	return OK;
-}
+	return OK;}
